minor specified cast to silence warning

+++ soul-sifter/NSString+MD5.mm
@@ -15,7 +15,7 @@
 - (NSString *)MD5 {
     const char *cstr = [self UTF8String];
     unsigned char result[16];
-    CC_MD5(cstr, strlen(cstr), result);
+    CC_MD5(cstr, (unsigned int)strlen(cstr), result);
     
     return [NSString stringWithFormat:
             @"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
